+++ R/print.model_set.R
@@ -46,6 +46,17 @@
 #' model list. Default is `"original"`.
 #' Used if `bpp_target` is not `NULL`.
 #'
+#' @param more_fit_measures Character
+#' vector. To be passed to
+#' [lavaan::fitMeasures()]. Default is
+#' `NULL`. If not `NULL`, these are
+#' the additional measures to be printed.
+#'
+#' @param fit_measures_digits The number of
+#' decimal places to be displayed
+#' for additional fit measures, if
+#' requested. Default is 3.
+#'
 #' @param ...  Optional arguments.
 #' Ignored.
 #'
@@ -81,6 +92,8 @@ print.model_set <- function(x,
                             max_models = 20,
                             bpp_target = NULL,
                             target_name = "original",
+                            more_fit_measures = NULL,
+                            fit_measures_digits = 3,
                             ...) {
     fit_n <- length(x$models)
     fit_names <- names(x$models)
@@ -126,6 +139,14 @@ print.model_set <- function(x,
                        format = "f")
         out_table["Cumulative"] <- tmp
       }
+    if (!is.null(more_fit_measures)) {
+        fit_fm <- sapply(x$fit,
+                         lavaan::fitMeasures,
+                         fit.measures = more_fit_measures,
+                         output = "vector")
+        fit_fm <- t(fit_fm)
+        out_table <- cbind(out_table, fit_fm)
+      }
     out_table_print <- out_table
     out_table_print$Prior <- round(out_table_print$Prior,
                                  digits = bpp_digits)
@@ -142,6 +163,14 @@ print.model_set <- function(x,
     out_table_print$BPP <- formatC(out_table_print$BPP,
                                    digits = bpp_digits,
                                    format = "f")
+    if (!is.null(more_fit_measures)) {
+        fm_names <- colnames(fit_fm)
+        for (xx in fm_names) {
+            out_table_print[xx] <- formatC(out_table_print[, xx, drop = TRUE],
+                                           digits = fit_measures_digits,
+                                           format = "f")
+          }
+      }
     if (isTRUE(fit_n > max_models)) {
         gt_max_models <- TRUE
         x_tmp <- out_table_print[seq_len(max_models), ]